+++ ChangeLog.md
@@ -1,7 +1,9 @@
 # Change Log #
-- There are now static and non-static versions of the decorators. XXXX
+- Appropriate decorators now return RMF::Vector3, RMF::Vector4 or
+	RMF::Vector3s instead of RMF::Floats.
+- There are now static and non-static versions of the decorators.
 - You now have to explicitly add a frame or set the current frame before
-	manipulating frame data. XXXX
+	manipulating frame data.
 - RMF::NodeConstHandle::get_value() subsumes get_value_always() which was
 	removed. If there is no value, then either a RMF::Nullable for
 	which RMF::Nullable::get_is_null() is true is returned (C++) or